% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nc.R
\name{create_nc_file_app}
\alias{create_nc_file_app}
\title{Request ERA5 file using an UI.}
\usage{
create_nc_file_app()
}
\description{
Create a request to ERA5 complete reanalysis database with a user interface.
}
\seealso{
\code{\link{create_nc_file_data}}
}
