% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nc.R
\name{create_nc_file_data}
\alias{create_nc_file_data}
\title{Request ERA5 file.}
\usage{
create_nc_file_data(
  levels,
  hours,
  grid,
  target,
  start_date,
  end_date,
  coordinates,
  variables,
  units_num
)
}
\arguments{
\item{levels}{levels id requested joined by comma ",". The range is from 1 to 127 both included}

\item{hours}{requested in HH:MM:SS format joined by comma ",".}

\item{grid}{of the longitude-latitude net in RxR format where R stands for a real number}

\item{target}{name of the .nc file. Must end with ".nc" extension.}

\item{start_date}{first day to request in format YYYY/MM/DD (included).}

\item{end_date}{last day to request in format YYYY/MM/DD (included).}

\item{coordinates}{string with the four coordinates joined by slash "/" (N/W/E/S)}

\item{variables}{integer representation of the variables to request joined by comma ",".}

\item{units_num}{number column of the unit of the level. 3 for half-level level, 5 for Geopotential altitude and 7 for Temperature}
}
\description{
Create a request to ERA5 complete reanalysis database and return a .nc
file with the data requested
}
\seealso{
\url{https://apps.ecmwf.int/codes/grib/param-db/}

\url{https://confluence.ecmwf.int/display/UDOC/L137+model+level+definitions}
}
\keyword{internal}
