% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_library.R
\name{fetch_information}
\alias{fetch_information}
\title{Fetch configuration information}
\usage{
fetch_information(nc_file)
}
\arguments{
\item{nc_file}{name of the .nc file.}
}
\value{
A list with the important information stored in the .nc file.
}
\description{
Fetch configuration information
}
