% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alter_nc.R
\name{get_merge_observed}
\alias{get_merge_observed}
\title{Merge with observed values}
\usage{
get_merge_observed(
  data_era,
  data_obs,
  name_col = paste(data_obs$station_info$name, data_obs$var_info$name, sep =
    data_era$sep)
)
}
\arguments{
\item{data_era}{"ERA5" list to add.}

\item{data_obs}{Observed variable to add.}

\item{name_col}{name of the column in the data frame.}
}
\value{
the new "ERA5" list with the observed value added.
}
\description{
Add to a "ERA5" daily file an observed "ECA" variable.
The observed time set could be reduced in order to fit the "ERA5" time set.
}
