% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_library.R
\name{get_subset_city}
\alias{get_subset_city}
\title{Subset grid city}
\usage{
get_subset_city(data_era, city_name, depth = 1)
}
\arguments{
\item{data_era}{"ERA5" list to modify.}

\item{city_name}{name of the city.}

\item{depth}{of the net to subset.}
}
\value{
the "ERA5" subset.
}
\description{
Function that returns a subset of a "ERA5" list around a location.
}
\seealso{
\code{\link{read_nc_file}}
}
