% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alter_nc.R
\name{get_subset_coordinates}
\alias{get_subset_coordinates}
\title{Subset grid by coordinate point}
\usage{
get_subset_coordinates(data_era, coordinate, depth = 1)
}
\arguments{
\item{data_era}{"ERA5" list to modify.}

\item{coordinate}{array with the longitude and latitude.}

\item{depth}{of the net to subset.}
}
\value{
the "ERA5" subset.
}
\description{
Function that returns a subset of a "ERA5" list around a coordinate.
}
\seealso{
\code{\link{read_nc_file}}
}
