% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates_functions.R
\name{lat_to_str}
\alias{lat_to_str}
\title{Latitude integer to string representation}
\usage{
lat_to_str(latitude)
}
\arguments{
\item{latitude}{latitude array in integer representation}
}
\value{
An array of strings with the string representation of latitude.
}
\description{
Latitude integer to string representation
}
\examples{
lat_to_str(c(1,-1))
}
