% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates_functions.R
\name{lon_to_str}
\alias{lon_to_str}
\title{Longitude integer to string representation}
\usage{
lon_to_str(longitude)
}
\arguments{
\item{longitude}{longitude array in integer representation}
}
\value{
an array of strings with the string representation of longitude
}
\description{
Longitude integer to string representation
}
\examples{
lon_to_str(c(1,-1))
}
