% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eca_library.R
\name{read_eca_zip}
\alias{read_eca_zip}
\title{Extract data from ECA&D database}
\usage{
read_eca_zip(zip_file, blended)
}
\arguments{
\item{zip_file}{name of the zip extracted from the ECA&D database.}

\item{blended}{TRUE if the data is blended, FALSE otherwise.}
}
\value{
a list with all the variables contained in the zip_file
}
\description{
Function that extract all the data downloaded from the ECA&D database.
}
\seealso{
\url{https://www.ecad.eu/dailydata/customquery.php}
}
