% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_library.R
\name{read_nc_file}
\alias{read_nc_file}
\title{Read data from .nc}
\usage{
read_nc_file(
  nc_file,
  reduction.x = 1L,
  reduction.y = 1L,
  sep = "_",
  hourly.col = TRUE,
  var.names = .var_names(nc_open(nc_file)),
  geo.levels = .geo_levels(nc_open(nc_file)),
  extended.hours = FALSE,
  display.times = FALSE
)
}
\arguments{
\item{nc_file}{name of the nc file}

\item{reduction.x}{integer reduction factor of the longitude grid.}

\item{reduction.y}{integer reduction factor of the latitude grid.}

\item{sep}{character to separate fields of the result .csv}

\item{hourly.col}{TRUE if the hours are meant to be in the columns.}

\item{var.names}{Name of the variables to save. Every variable as default.}

\item{geo.levels}{Levels to save. Every level as default.}

\item{extended.hours}{TRUE to display the hours with minutes and seconds.}

\item{display.times}{TRUE to display a time-report of the function.}
}
\value{
a list with the information in the .nc file
}
\description{
Create a "ERA5" list containing the data frame with the information stored in a .nc file.
}
