% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates_functions.R
\name{str_to_lat}
\alias{str_to_lat}
\title{String latitude representation to integer value}
\usage{
str_to_lat(lat_str)
}
\arguments{
\item{lat_str}{string array representing the latitudes.}
}
\value{
an integer array with the integer representation
}
\description{
String latitude representation to integer value
}
\examples{
str_to_lat(c("1N","1S"))
}
