% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates_functions.R
\name{str_to_lon}
\alias{str_to_lon}
\title{String longitude representation to integer value}
\usage{
str_to_lon(lon_str)
}
\arguments{
\item{lon_str}{string array representing the longitudes.}
}
\value{
an integer array with the integer representation
}
\description{
String longitude representation to integer value
}
\examples{
str_to_lon(c("1E","1W"))
}
