% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates_functions.R
\name{transform_coordinate}
\alias{transform_coordinate}
\title{Transform (0,360) coordinate format to (-180,180)}
\usage{
transform_coordinate(v)
}
\arguments{
\item{v}{array of latitudes or longitudes}
}
\value{
array with the transformed longitudes or latitudes
}
\description{
Transform (0,360) coordinate format to (-180,180)
}
\examples{
transform_coordinate(c(360,0))
}
