% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_library.R
\name{write_csv}
\alias{write_csv}
\title{Writing .csv of .nc file}
\usage{
write_csv(
  nc_file,
  reduction.x = 1L,
  reduction.y = 1L,
  prefix = "",
  sep = "_",
  root = ".",
  hourly.col = TRUE,
  var.names = .var_names(nc_open(nc_file)),
  geo.levels = .geo_levels(nc_open(nc_file)),
  extended.hours = FALSE,
  display.times = FALSE
)
}
\arguments{
\item{nc_file}{name of the .nc file}

\item{reduction.x}{integer reduction factor of the longitude grid.}

\item{reduction.y}{integer reduction factor of the latitude grid.}

\item{prefix}{of the column names in the data frame.}

\item{sep}{character to separate fields of the result .csv}

\item{root}{to set the create directory.}

\item{hourly.col}{TRUE if the hours are meant to be in the columns.}

\item{var.names}{Name of the variables to save. Every variable as default.}

\item{geo.levels}{Levels to save. Every level as default.}

\item{extended.hours}{TRUE to display the hours with minutes and seconds.}

\item{display.times}{TRUE to display a time-report of the function.}
}
\description{
Create a .csv file with the information stored in a .nc file.
}
