% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_library.R
\name{fetch_information}
\alias{fetch_information}
\title{Information of a .nc.}
\usage{
fetch_information(nc_file, reduction.x = 1L, reduction.y = 1L)
}
\arguments{
\item{nc_file}{.nc file.}

\item{reduction.x}{integer reduction factor of the longitude grid.}

\item{reduction.y}{integer reduction factor of the latitude grid.}
}
\value{
a list with the information of the .nc file.
}
\description{
Fetch the most important information about a .nc file.
}
\seealso{
\code{\link{read_nc_file}}
}
