% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_library.R
\name{read_nc_file}
\alias{read_nc_file}
\title{Read a .nc file}
\usage{
read_nc_file(configuration)
}
\arguments{
\item{configuration}{of the request}
}
\value{
a data-frame with the information of the request
}
\description{
Read a .nc file
}
